Move the rest of the library_utils
