+++ crates/edr_solidity/src/library_utils.rs
@@ -1,7 +1,25 @@
 //! Utility functions for working with libraries in Solidity.
 //! Ported from `hardhat-network/stack-traces/library-utils.ts`.
 
+use edr_evm::hex;
+
 use crate::artifacts::CompilerOutputBytecode;
+
+/// Normalizes the compiler output bytecode by replacing the library addresses
+/// with zeros.
+pub fn normalize_compiler_output_bytecode(
+    mut compiler_output_bytecode_object: String,
+    addresses_positions: &[u32],
+) -> Result<Vec<u8>, hex::FromHexError> {
+    const ZERO_ADDRESS: &str = "0000000000000000000000000000000000000000";
+
+    for &pos in addresses_positions {
+        compiler_output_bytecode_object =
+            link_hex_string_bytecode(compiler_output_bytecode_object, ZERO_ADDRESS, pos);
+    }
+
+    hex::decode(compiler_output_bytecode_object)
+}
 
 /// Retrieves the positions of the library addresses in the bytecode.
 pub fn get_library_address_positions(bytecode_output: &CompilerOutputBytecode) -> Vec<u32> {
